Tyuiu.IbrayevAA.Sprint2.Task6.V14.Lib/DataService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using tyuiu.cources.programming.interfaces.Sprint2;

namespace Tyuiu.IbrayevAA.Sprint2.Task6.V14.Lib
{
    public class DataService : ISprint2Task6V14
    {
        public string FindDayName(int k, int d)
        {

            string res;
            switch ((k + d - 2) % 7 + 1)
            {
                case 1: return "понедельник";
                case 2: return "вторник";
                case 3: return "среда";
                case 4: return "четверг";
                case 5: return "пятница";
                case 6: return "суббота";
                case 7: return "воскресенье";
                default: throw new ArgumentException($"Число k должно быть от 1 до 365 и число d должно быть от 1 до 7. Значение {k} {d}");
            }
        }
    }
}
